feat:用户服务代码

+++ src/main/java/com/card/mapper/UserMapper.java
@@ -13,7 +13,7 @@ public interface UserMapper {
 
     int deleteByPrimaryKey(Integer id);
 
-    int insert(User record);
+    long insert(User record);
 
     User selectByPrimaryKey(Integer id);
